git-svn-id: file:///tmp/svnroot_only_rec/internal/rec-project/rec/trunk@4676 8c4c7310-a39f-45c4-bd0d-b10b2e5d99a4

+++ rec/src/java/common/com/linkare/rec/data/acquisition/TOTAL_SAMPLES_UNDEFINED.java
@@ -1,5 +1,5 @@
 package com.linkare.rec.data.acquisition;
 
 public interface TOTAL_SAMPLES_UNDEFINED {
-	public static final int value = (int) ((int) -1);
+	public static final int value = (-1);
 }
